fixed logic in file_repository.py

+++ backend/app/repositories/file_repository.py
@@ -1,11 +1,15 @@
 import os
-from fastapi import UploadFile
+from fastapi import UploadFile, HTTPException
 
 
 async def create_file(base_dir: str, user_id: str, filename: str, file: UploadFile.file):
+
     folder_path = os.path.join(base_dir, user_id)
     if os.path.exists(folder_path):
         os.mkdir(folder_path)
+
+    if check_file_exists_by_user(base_dir, user_id, file.filename):
+        raise HTTPException(status_code=400)
 
     file_path = os.path.join(folder_path, filename)
     with open(file_path, "wb") as f:
@@ -16,4 +20,5 @@ async def create_file(base_dir: str, user_id: str, filename: str, file: UploadFi
 
 async def check_file_exists_by_user(base_dir: str, user_id: str, filename: str):
     folder_path = os.path.join(base_dir, user_id)
-    return os.path.exists(folder_path)
+    file_path = os.path.join(folder_path, filename)
+    return os.path.exists(file_path)
